Enum wasmer_clif_fork_wasm::GlobalVariable

pub enum GlobalVariable {
    Const(Value),
    Memory {
        gv: GlobalValue,
        offset: Offset32,
        ty: Type,
    },
}

The value of a WebAssembly global variable.

Variants

Const(Value)

This is a constant global with a value known at compile time.

Memory

This is a variable in memory that should be referenced through a GlobalValue.

Fields of Memory

gv: GlobalValue

The address of the global variable storage.

offset: Offset32

An offset to add to the address.

ty: Type

The global variable's type.