Struct wasmer_clif_fork_wasm::TranslationState

pub struct TranslationState {
    pub stack: Vec<Value>,
    pub control_stack: Vec<ControlStackFrame>,
    pub reachable: bool,
    // some fields omitted
}

Contains information passed along during the translation and that records:

The current value and control stacks.

The depth of the two unreachable control blocks stacks, that are manipulated when translating unreachable code;

Fields
stack: Vec<Value>

A stack of values corresponding to the active values in the input wasm function at this point.

control_stack: Vec<ControlStackFrame>

A stack of active control flow operations at this point in the input wasm function.

reachable: bool

Is the current translation state still reachable? This is false when translating operators like End, Return, or Unreachable.

Methods

impl TranslationState

pub fn new() -> Self

Construct a new, empty, TranslationState

pub fn initialize(&mut self, sig: &Signature, exit_block: Ebb)

Initialize the state for compiling a function with the given signature.

This resets the state to containing only a single block representing the whole function. The exit block is the last block in the function which will contain the return instruction.

pub fn push1(&mut self, val: Value)

Push a value.

pub fn pushn(&mut self, vals: &[Value])

Push multiple values.

pub fn pop1(&mut self) -> Value

Pop one value.

pub fn peek1(&self) -> Value

Peek at the top of the stack without popping it.

pub fn pop2(&mut self) -> (Value, Value)

Pop two values. Return them in the order they were pushed.

pub fn pop3(&mut self) -> (Value, Value, Value)

Pop three values. Return them in the order they were pushed.

pub fn popn(&mut self, n: usize)

Pop the top n values on the stack.

The popped values are not returned. Use peekn to look at them before popping.

pub fn peekn(&self, n: usize) -> &[Value]

Peek at the top n values on the stack in the order they were pushed.

pub fn push_block(&mut self, following_code: Ebb, num_result_types: usize)

Push a block on the control stack.

pub fn push_loop(
&mut self,
header: Ebb,
following_code: Ebb,
num_result_types: usize
)

Push a loop on the control stack.

pub fn push_if(
&mut self,
branch_inst: Inst,
following_code: Ebb,
num_result_types: usize
)

Push an if on the control stack.

impl TranslationState

Methods for handling entity references.

pub fn get_global<FE: FuncEnvironment + ?Sized>(
&mut self,
func: &mut Function,
index: u32,
environ: &mut FE
) -> WasmResult<GlobalVariable>

Get the GlobalVariable reference that should be used to access the global variable index. Create the reference if necessary. Also return the WebAssembly type of the global.

pub fn get_heap<FE: FuncEnvironment + ?Sized>(
&mut self,
func: &mut Function,
index: u32,
environ: &mut FE
) -> WasmResult<Heap>

Get the Heap reference that should be used to access linear memory index. Create the reference if necessary.

pub fn get_table<FE: FuncEnvironment + ?Sized>(
&mut self,
func: &mut Function,
index: u32,
environ: &mut FE
) -> WasmResult<Table>

Get the Table reference that should be used to access table index. Create the reference if necessary.

pub fn get_indirect_sig<FE: FuncEnvironment + ?Sized>(
&mut self,
func: &mut Function,
index: u32,
environ: &mut FE
) -> WasmResult<(SigRef, usize)>

Get the SigRef reference that should be used to make an indirect call with signature index. Also return the number of WebAssembly arguments in the signature.

Create the signature if necessary.

pub fn get_direct_func<FE: FuncEnvironment + ?Sized>(
&mut self,
func: &mut Function,
index: u32,
environ: &mut FE
) -> WasmResult<(FuncRef, usize)>

Get the FuncRef reference that should be used to make a direct call to function index. Also return the number of WebAssembly arguments in the signature.

Create the function reference if necessary.